Update configuration for bspwm, sxhkd, and vim
Add new rules for `bspwm`.

Remove pointer related bindings in `sxhkd` which are no longer needed.

Be more picky about what to include in the file selection list for
`CtrlP` and `wildmenu` in `vim`.

Enable the `matchit` plugin for `vim`.

Use a different `sxhkd` syntax plugin for `vim`.

Update the `minpac` submodule.