Add support for `zsh`
Add `zsh` module, which includes a `.zshrc`, some aliases, and a
`powerlevel10k` theme configuration.

Make use of the theme feature in `rofi.

Remove the `-f` flag from defining the `$VISUAL` editor. The `-f` flag
was not treated as a flag, but rather as part of the name, causing the
system to look for a single file named `gvim -f` to launch.